backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn==0.29.0

# LLM
anthropic==0.28.0

# Web scraping
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.32.2
lxml==5.2.2
selenium==4.21.0

# Search
tavily-python==0.3.3

# Data
pydantic==2.7.1
python-dotenv==1.0.1

# Async
httpx==0.27.0
